release-manager-v2/orquestador-core/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orquestador-core"
version = "0.1.0"
dependencies = [
    "python-dotenv",
    "requests",
    "PyGithub",
    "jira",
    "GitPython",
    "typer"
]
